Rename parameter with typo
While reading through render.c I noticed a small typo in one of the parameter names. I assume that udaa was intended to be udata. Since the parameter is not actually used in transform_smooth at present, I have renamed it to _ as done in some of the other transform functions.